+++ src/main/java/br/com/ffroliva/portfolio/model/UserPrincipal.java
@@ -18,7 +18,7 @@ public class  UserPrincipal implements UserDetails {
 
 	private static final long serialVersionUID = -1111403067001073490L;
 
-	private Long id;
+	private Integer id;
 
     private String name;
 
@@ -32,7 +32,7 @@ public class  UserPrincipal implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserPrincipal(Long id, String name, String username, String email, String password, Collection<? extends GrantedAuthority> authorities) {
+    public UserPrincipal(Integer id, String name, String username, String email, String password, Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.name = name;
         this.username = username;
